Remove server grab around EShapeSetShape().
If the server grab is removed from here, it would seem appropriate
to add it in EShapeSetShape() (or EShapeCombineShape()).
However, I'm not sure we need atomicity at this point.
If we read an "incorrrect" shape, we should have an unprocessed shape
change notification pending.
Let's just see if this causes trouble.

SVN revision: 70399