Use initially created GC for various operations.
A general "root window GC" is fine for most operations needing one.
Using this we can avoid some GC caching and in certain situations
numerous GC create/use/destroy sequences.

As part of this the default GC subwindow_mode is changed to
IncludeInferiors, which reduces the number of places we need a
non-default GC.