Fix focus issues with pager zoom windows on desks other than first (T2197).
Change the zoom window parent to be the current desktop in stead of the
root one.
Otherwise, when entering the zoom window (on desktops other than first),
we get an indication that we are leaving the current desktop causing the
current active window to be unfocused (and refocused when leaving).